Refactoring test
